+++ packages/core/src/actions/index.ts
@@ -1,6 +1,7 @@
 /* Casino */
 
 export * from "../read/common/tokenMetadata";
+export * from "./api/freebet/claim";
 export * from "./api/leaderboard/leaderboards";
 export * from "./casino/cointoss";
 export * from "./casino/dice";
@@ -13,5 +14,7 @@ export * from "./casino/weightedGame";
 export * from "./casino/wheel";
 /* Common */
 export * from "./common/approve";
+/* Freebet */
+export * from "./freebet/claim";
 /* Leaderboard */
 export * from "./leaderboard/leaderboard";