+++ packages/core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @betswirl/sdk-core
 
+## 0.1.26
+
+### Patch Changes
+
+- Fix claim freebet code parsing
+
+## 0.1.25
+
+### Patch Changes
+
+- Fix claimFreebetCode (number => toString())
+
+## 0.1.24
+
+### Patch Changes
+
+- Fix claim freebet code APi call body
+
+## 0.1.23
+
+### Patch Changes
+
+- Add sign/claim freebet code
+
 ## 0.1.22
 
 ### Patch Changes